- Added test for TableViewKitViewController. - Split TableViewKitTest into two Unit Test classes due to the SwiftLint File Length Rule.
